Moved processing less relevant to X.509 to a separate module.
